backend/requirements.txt
# Core API
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==44.0.0

# Async HTTP
httpx==0.28.1
aiofiles==24.1.0

# Job Queue
celery==5.4.0
redis==5.2.1
flower==2.0.1

# AI
anthropic==0.43.0

# Browser Automation
playwright==1.49.1

# Fuzzy matching
thefuzz==0.22.1
python-Levenshtein==0.26.1

# Utilities
python-dotenv==1.0.1
email-validator==2.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.1
httpx==0.28.1
